Implemented death screen, victory screen, and estus on 2nd and 3rd levels

+++ Assets/Scripts/Managers/CharacterStats.cs
@@ -6,11 +6,11 @@ public class CharacterStats : MonoBehaviour
 {
 
     public int healthLevel = 10;
-    public int staminaLevel = 10;
+    public float staminaLevel = 10f;
     public int maxHealth;
     public int currentHealth;
-    public int maxStamina;
-    public int currentStamina;
+    public float maxStamina;
+    public float currentStamina;
 
 
     public bool isDead;
