+++ Assets/Scripts/UI/StaminaBar.cs
@@ -8,13 +8,13 @@ public class StaminaBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxStamina(int maxStamina)
+    public void SetMaxStamina(float maxStamina)
     {
         slider.maxValue = maxStamina;
         slider.value = maxStamina;
     }
 
-    public void SetCurrentStamina(int currentStamina)
+    public void SetCurrentStamina(float currentStamina)
     {
         slider.value = currentStamina;
     }
